added new start level including some work for inner black areas

+++ Dungeon-FrontEnd-GraphicsEngine2D/src/main/java/graphics/util/RoomSize.java
@@ -3,7 +3,7 @@ package graphics.util;
 public class RoomSize {
 	
 	public static int by(int divisor, int roomSize) {
-		return  Math.round(roomSize / divisor);
+		return  (int) Math.round(((double)roomSize) / divisor);
 	}
 
 }
